Revert "Bug 2015410 - Part 2: Fix clang-tidy warnings in js/src/builtin/temporal. r=spidermonkey-reviewers,arai" for causing Calendar.cpp related failures.
This reverts commit 1f9af96e8664f756c4238384a3714a7295e2a55b.

This reverts commit 9d0bbd803fcdebd0650f3657a3d328940a6ca47a.

UltraBlame original commit: d270382b770b3298da5b7fe796cd16982af23f74

+++ js/src/builtin/temporal/TimeZone.cpp
@@ -1064,8 +1064,7 @@ bool js::temporal::DisambiguatePossibleEpochNanoseconds(
                "subtracting nanoseconds is at most one day");
 
     
-    auto earlierDate = BalanceISODate(isoDateTime.date,
-                                      static_cast<int32_t>(earlierTime.days));
+    auto earlierDate = BalanceISODate(isoDateTime.date, earlierTime.days);
 
     
     auto earlierDateTime = ISODateTime{earlierDate, earlierTime.time};
@@ -1095,8 +1094,7 @@ bool js::temporal::DisambiguatePossibleEpochNanoseconds(
              "adding nanoseconds is at most one day");
 
   
-  auto laterDate =
-      BalanceISODate(isoDateTime.date, static_cast<int32_t>(laterTime.days));
+  auto laterDate = BalanceISODate(isoDateTime.date, laterTime.days);
 
   
   auto laterDateTime = ISODateTime{laterDate, laterTime.time};
